# schemas/openapi/odyssey-v1.openapi.json
{
  "components": {
    "schemas": {
...
      "AccessTokenRefreshRequest": {
        "additionalProperties": false,
        "properties": {
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "refresh_credential": {
            "maxLength": 512,
            "minLength": 32,
            "title": "Refresh Credential",
            "type": "string"
          }
        },
        "required": [
          "device_id",
          "refresh_credential"
        ],
        "title": "AccessTokenRefreshRequest",
        "type": "object"
      },
      "AccessTokenResponse": {
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "access_token_expires_at": {
            "format": "date-time",
            "title": "Access Token Expires At",
            "type": "string"
          },
          "token_type": {
            "const": "Bearer",
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "access_token_expires_at"
        ],
        "title": "AccessTokenResponse",
        "type": "object"
      },
      "AppleChallengeRequest": {
        "additionalProperties": false,
        "properties": {
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          }
        },
        "required": [
          "device_id"
        ],
        "title": "AppleChallengeRequest",
        "type": "object"
      },
      "AppleChallengeResponse": {
        "additionalProperties": false,
        "properties": {
          "challenge_id": {
            "format": "uuid",
            "title": "Challenge Id",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "nonce": {
            "title": "Nonce",
            "type": "string"
          }
        },
        "required": [
          "challenge_id",
          "nonce",
          "expires_at"
        ],
        "title": "AppleChallengeResponse",
        "type": "object"
      },
      "AppleExchangeRequest": {
        "additionalProperties": false,
        "properties": {
          "app_version": {
            "maxLength": 100,
            "minLength": 1,
            "title": "App Version",
            "type": "string"
          },
          "challenge_id": {
            "format": "uuid",
            "title": "Challenge Id",
            "type": "string"
          },
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "display_name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "identity_token": {
            "maxLength": 16384,
            "minLength": 100,
            "title": "Identity Token",
            "type": "string"
          },
          "nonce": {
            "maxLength": 200,
            "minLength": 32,
            "title": "Nonce",
            "type": "string"
          },
          "platform": {
            "$ref": "#/components/schemas/DevicePlatform"
          }
        },
        "required": [
          "challenge_id",
          "device_id",
          "nonce",
          "identity_token",
          "display_name",
          "platform",
          "app_version"
        ],
        "title": "AppleExchangeRequest",
        "type": "object"
      },
...
      "DeviceEnrollmentResponse": {
        "additionalProperties": false,
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "access_token_expires_at": {
            "format": "date-time",
            "title": "Access Token Expires At",
            "type": "string"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceSummary"
          },
          "refresh_credential": {
            "title": "Refresh Credential",
            "type": "string"
          },
          "refresh_credential_expires_at": {
            "format": "date-time",
            "title": "Refresh Credential Expires At",
            "type": "string"
          },
          "token_type": {
            "const": "Bearer",
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "access_token_expires_at",
          "refresh_credential",
          "refresh_credential_expires_at",
          "device"
        ],
        "title": "DeviceEnrollmentResponse",
        "type": "object"
      },
      "DeviceListResponse": {
        "additionalProperties": false,
        "properties": {
          "devices": {
            "items": {
              "$ref": "#/components/schemas/DeviceSummary"
            },
            "title": "Devices",
            "type": "array"
          }
        },
        "required": [
          "devices"
        ],
        "title": "DeviceListResponse",
        "type": "object"
      },
      "DevicePlatform": {
        "enum": [
          "ios",
          "ipados",
          "macos",
          "watchos",
          "visionos"
        ],
        "title": "DevicePlatform",
        "type": "string"
      },
      "DeviceRevocationReason": {
        "enum": [
          "compromised",
          "lost",
          "owner_request",
          "reinstalled",
          "retired"
        ],
        "title": "DeviceRevocationReason",
        "type": "string"
      },
      "DeviceRevocationRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "$ref": "#/components/schemas/DeviceRevocationReason"
          }
        },
        "required": [
          "reason"
        ],
        "title": "DeviceRevocationRequest",
        "type": "object"
      },
      "DeviceStatus": {
        "enum": [
          "active",
          "revoked"
        ],
        "title": "DeviceStatus",
        "type": "string"
      },
      "DeviceSummary": {
        "additionalProperties": false,
        "properties": {
          "app_version": {
            "title": "App Version",
            "type": "string"
          },
          "device_id": {
            "format": "uuid",
            "title": "Device Id",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "enrolled_at": {
            "format": "date-time",
            "title": "Enrolled At",
            "type": "string"
          },
          "last_authenticated_at": {
            "format": "date-time",
            "title": "Last Authenticated At",
            "type": "string"
          },
          "last_seen_at": {
            "format": "date-time",
            "title": "Last Seen At",
            "type": "string"
          },
          "platform": {
            "$ref": "#/components/schemas/DevicePlatform"
          },
          "revocation_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceRevocationReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "revoked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "status": {
            "$ref": "#/components/schemas/DeviceStatus"
          }
        },
        "required": [
          "device_id",
          "display_name",
          "platform",
          "app_version",
          "status",
          "enrolled_at",
          "last_authenticated_at",
          "last_seen_at"
        ],
        "title": "DeviceSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Private local-first personal navigation API",
    "title": "Odyssey API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/apple/challenges": {
      "post": {
        "operationId": "create_apple_challenge_v1_auth_apple_challenges_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleChallengeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppleChallengeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Apple Challenge",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/apple/exchange": {
      "post": {
        "operationId": "exchange_apple_identity_v1_auth_apple_exchange_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceEnrollmentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Exchange Apple Identity",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/devices": {
      "get": {
        "operationId": "list_devices_v1_auth_devices_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Devices",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/devices/{device_id}/revoke": {
      "post": {
        "operationId": "revoke_device_v1_auth_devices__device_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRevocationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Revoke Device",
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/token/refresh": {
      "post": {
        "operationId": "refresh_access_token_v1_auth_token_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessTokenRefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Access Token",
        "tags": [
          "authentication"
        ]
      }
    },
...
  }
}